销售报价单导出

+++ template/sales_quotation/create.tpl
@@ -25,20 +25,23 @@
         <!-- Main content -->
         <section class="content">
             <!-- Default box -->
+            <form class="form-horizontal" action="/sales_quotation/add_sales_quotation.php" method="post">
             <div class="box">
                 <div class="box-header with-border">
 
                     <div class="box-body">
                         
                         <span class="col-md-5">报价单名称 <input type="text" name="sales_quotation_name" value=""></span>
-                        <span class="col-md-4">客户名称    <select>
-                                        <option>请选择</option>
+                        <span class="col-md-3">客户名称 
+                                    <select name="customer_id">
+                                        <option value="0">请选择</option>
                                         <{foreach from=$listCustomer item=item}>
-                                            <option value="customer_id"><{$item.customer_name}></option>
+                                            <option value="<{$item.customer_id}>"><{$item.customer_name}></option>
                                         <{/foreach}>
                                    </select>
                        </span>
-                       <span class="col-md-3">加价规则   <input type="text" name="plue_price" size=3 name="plue_price"></span>
+                       <span class="col-md-4">加价规则   <input type="text" name="plue_price" size=3 value="<{$smarty.session.plue_price}>" id="plue_price">
+                            <a href="javascript:void(0);" class="btn btn-primary btn-sm" id="refresh" style="margin-left: 10px;"><i class="fa fa-refresh"></i> 刷新工费</a></span>
                     </div>
                 
                 </div>
@@ -46,7 +49,7 @@
             </div>
             <div class="box">
                 <div class="box-header with-border">
-                    <span class="col-md-2"><input type="checkbox" name="select-all"> 全选</span>
+                    <span class="col-md-2"><input type="checkbox" name="check-all"> 全选</span>
                     <span class="col-md-2"><a href="javascript:void(0);" class="btn btn-primary btn-sm" id="delMulti" style="margin-left: 10px;"><i class="fa fa-trash-o"></i> 批量删除</a></span>
                     <span class="col-md-2"><i class="fa fa-shopping-cart">共计<{if $countSpu!=""}><{$countSpu}><{else}>0<{/if}>款商品</i></span>
            
@@ -57,8 +60,8 @@
                             <thead>
                                 <tr>
                                     <th rowspan="2">选择</th>
-                                    <th rowspan="2">SKU编号</th>
-                                    <th rowspan="2">SKU名称</th>
+                                    <th rowspan="2">SPU编号</th>
+                                    <th rowspan="2">SPU名称</th>
                                     <th rowspan="2">产品图片</th>
                                     <th rowspan="2">三级分类</th>
                                     <th rowspan="2">规格重量(g)</th>
@@ -74,14 +77,14 @@
                                     <th>K黄</th>
                                     <th>红白</th>
                                     <th>红黄</th>
-                                    <th>白黄</th>
+                                    <th>黄白</th>
                                     <th>三色</th>
                                 </tr>
                             </thead>
                             <tbody>
 <{foreach from=$listSpuInfo item=item name=foo}>                               
                                 <tr>
-                                    <td><input type="checkbox" name="select-spu"></td>
+                                    <td><input type="checkbox" name="spu_id[]" value="<{$item.spu_id}>" /></td>
                                     <td><{$item.spu_id}></td>
                                     <td><{$item.spu_name}></td>
                                     <td><img src="<{$item.image_url}>" class="width-100" alt="..."></td>
@@ -89,23 +92,30 @@
                                     <td><{$item.weight_value}></td>
                                     <td><{$item.size_name}></td>
                                     <td><{$item.material_name}></td>
-                                    <td><input type="text" size="4" value="<{$item['K红']}>"></td>
-                                    <td><input type="text" size="4" value="<{$item['K白']}>"></td>
-                                    <td><input type="text" size="4" value="<{$item['K黄']}>"></td>
-                                    <td><input type="text" size="4" value="<{$item['红白']}>"></td>
-                                    <td><input type="text" size="4" value="<{$item['红黄']}>"></td>
-                                    <td><input type="text" size="4" value="<{$item['白黄']}>"></td>
-                                    <td><input type="text" size="4" value="<{$item['三色']}>"></td>
+                                    <td><input type="text" size="4" name="<{$item.spu_id}>[<{$listIndexColorName['K红']}>]" value="<{$item['K红']}>"></td>
+                                    <td><input type="text" size="4" name="<{$item.spu_id}>[<{$listIndexColorName['K白']}>]" value="<{$item['K白']}>"></td>
+                                    <td><input type="text" size="4" name="<{$item.spu_id}>[<{$listIndexColorName['K黄']}>]" value="<{$item['K黄']}>"></td>
+                                    <td><input type="text" size="4" name="<{$item.spu_id}>[<{$listIndexColorName['红白']}>]" value="<{$item['红白']}>"></td>
+                                    <td><input type="text" size="4" name="<{$item.spu_id}>[<{$listIndexColorName['红黄']}>]" value="<{$item['红黄']}>"></td>
+                                    <td><input type="text" size="4" name="<{$item.spu_id}>[<{$listIndexColorName['黄白']}>]" value="<{$item['黄白']}>"></td>
+                                    <td><input type="text" size="4" name="<{$item.spu_id}>[<{$listIndexColorName['三色']}>]" value="<{$item['三色']}>"></td>
                                     <td><{$item.spu_remark}></td>
-                                    <td><a href="/sales_quotation/cart_spu_delete.php?spu_id=<{$item.spu_id}>"><i class="fa fa-trash-o"></i></a></td>
+                                    <td><a href="/sales_quotation/cart_spu_delete.php?spu_id=<{$item.spu_id}>" class="delete-confirm"><i class="fa fa-trash-o"></i></a></td>
                                 </tr>
 <{/foreach}>                                
                             </tbody>
-                        </table>
+                        </table> 
                     </div>
+                    <div class="form-group">
+                        <div class="col-sm-offset-2">
+                            <button type="submit" class="btn btn-primary pull-right"><i class="fa fa-save"></i> 提交报价单</button>
+                        </div>
+                    </div>
+                    
                 </div>
                 <!-- /.box-body -->
             </div>
+            </form>
             <!-- /.box -->
         </section>
         <!-- /.content -->
@@ -130,5 +140,58 @@
 <!-- ./wrapper -->
 
 <{include file="section/foot.tpl"}>
+<script>
+$(function(){
+    
+    $('input[name="check-all"]').click(function () {
+
+        $('input[name="spu_id[]"]').prop('checked', $(this).prop('checked'));
+    });
+  
+
+    $('#delMulti').click(function(){
+
+        var chk_value =[]; 
+        $('input[name="spu_id[]"]:checked').each(function(){ 
+        
+            chk_value.push($(this).val()); 
+        });
+        
+        if(chk_value.length==0){
+            
+            alert("请选择SPU");
+            
+            return false; 
+        }
+
+        $.post('/sales_quotation/ajax_cart_spu_delete.php', {
+            spu_id              : chk_value,
+            '__output_format'   : 'JSON'
+        }, function (response) {
+
+            if (0 != response.code) {
+
+                showMessage('错误', response.message);
+
+                return  ;
+            }else{
+            
+                history.go(0);
+            }
+            
+        }, 'json');  
+    })
+
+    $('.delete-confirm').click(function () {
+
+        return  confirm('确认删除？');
+    });
+    $('#refresh').click(function(){
+    
+        pluePrice = $("#plue_price").val();
+        location.href = "/sales_quotation/create.php?plue_price="+pluePrice;
+    });
+})    
+</script>
 </body>
 </html>
